refactor: the monolithic xaml-converter.ts file into focused modules with clear separation of concerns.
- Added XamlElementProcessor for handling various XAML elements including paragraphs, runs, spans, lists, tables, and hyperlinks.
- Introduced XamlFormatter for text formatting and styling operations, including inline formatting and heading detection.
- Created XamlListProcessor to manage nested lists and maintain proper indentation in markdown.
- Developed XamlParser for XML parsing, content cleaning, and markdown normalization.
- Implemented XamlUtils for utility functions related to XAML element validation and markdown syntax detection.
- Enhanced overall structure and organization of XAML processing modules for improved maintainability and readability.
