+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- 2025-07-30: Refactor monolithic xaml-converter.ts into focused modules reducing complexity from 1,181 lines to 4 specialized modules
 - Release v1.0.3: Initial beta release
 - Release v1.0.4: correct Windows database search path from Logos4 to Logos directory
 - Release v1.0.5: add SHA256 checksum files for binary integrity verification
