+++ packages/core/src/types/xaml-types.ts
@@ -1,0 +1,51 @@
+// Shared TypeScript interfaces and types for XAML conversion modules
+// Defines common data structures used across XAML parser, processor, formatter, and utilities
+// Maintains type safety and consistency between refactored modules
+
+export interface XamlConverterOptions {
+  /** Font sizes that correspond to heading levels [H1, H2, H3, H4, H5, H6] */
+  headingSizes: number[];
+  /** Font family name used to identify code elements */
+  monospaceFontName: string;
+  /** Whether to ignore unknown elements */
+  ignoreUnknownElements: boolean;
+  /** Use HTML sub/superscript tags instead of Pandoc-style formatting */
+  htmlSubSuperscript: boolean;
+  /** Output directory for the export (used for images subfolder) */
+  outputDirectory?: string;
+  /** Current note filename for generating unique image names */
+  noteFilename?: string;
+  /** Whether to download images locally */
+  downloadImages: boolean;
+  /** Maximum image download size in MB */
+  maxImageSizeMB: number;
+  /** Download timeout in milliseconds */
+  downloadTimeoutMs: number;
+  /** Number of download retry attempts */
+  downloadRetries: number;
+  /** Convert all indent levels to blockquotes instead of &nbsp; indentation */
+  convertIndentsToQuotes: boolean;
+  /** Optional logging callback for verbose output */
+  onLog?: (message: string) => void;
+  /** Whether to enable verbose logging */
+  verbose?: boolean;
+}
+
+export interface XamlElement {
+  // Keep this as `any` because we don't know what the XAML elements will be
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  [key: string]: any;
+  '@_FontSize'?: string;
+  '@_FontWeight'?: string;
+  '@_FontStyle'?: string;
+  '@_FontFamily'?: string;
+  '@_BorderThickness'?: string;
+  '@_BorderBrush'?: string;
+  '@_Text'?: string;
+  '@_Tag'?: string;
+  '@_NavigateUri'?: string;
+  '@_MarkerStyle'?: string;
+  '@_Kind'?: string; // Added for list Kind
+  '@_Margin'?: string; // Added for paragraph indentation
+  '#text'?: string;
+}
