+++ packages/core/src/xaml-lists-processor.ts
@@ -1,21 +1,5 @@
-// XamlElement interface - copied from main converter for type safety
-interface XamlElement {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [key: string]: any;
-  '@_FontSize'?: string;
-  '@_FontWeight'?: string;
-  '@_FontStyle'?: string;
-  '@_FontFamily'?: string;
-  '@_BorderThickness'?: string;
-  '@_BorderBrush'?: string;
-  '@_Text'?: string;
-  '@_Tag'?: string;
-  '@_NavigateUri'?: string;
-  '@_MarkerStyle'?: string;
-  '@_Kind'?: string; // Added for list Kind
-  '@_Margin'?: string; // Added for paragraph indentation
-  '#text'?: string;
-}
+import type { XamlElement } from './types/xaml-types.js';
+import type { IXamlElementProcessor } from './types/processor-interface.js';
 
 /**
  * Context information for processing nested lists
@@ -41,10 +25,10 @@ interface ItemContent {
  * nested lists while maintaining proper indentation and markdown formatting.
  */
 export class XamlListProcessor {
-  private converter: any; // Reference to main XamlToMarkdownConverter
+  private processor: IXamlElementProcessor;
 
-  constructor(converter: any) {
-    this.converter = converter;
+  constructor(processor: IXamlElementProcessor) {
+    this.processor = processor;
   }
 
   /**
@@ -59,7 +43,7 @@ export class XamlListProcessor {
       if (!list) continue;
 
       // Extract list attributes using the converter's helper
-      const attrs = this.converter.getAttributes(list);
+      const attrs = this.processor.getAttributes(list);
       const markerStyle = attrs['@_Kind'] || attrs['@_MarkerStyle'] || 'Disc';
       const listType: 'ordered' | 'unordered' = markerStyle.toLowerCase() === 'decimal' ? 'ordered' : 'unordered';
       
@@ -130,7 +114,7 @@ export class XamlListProcessor {
     if (nestedLists.length > 0) {
       for (const nestedList of nestedLists) {
         // Extract list attributes to determine the list type for nested items
-        const attrs = this.converter.getAttributes(nestedList);
+        const attrs = this.processor.getAttributes(nestedList);
         const markerStyle = attrs['@_Kind'] || attrs['@_MarkerStyle'] || 'Disc';
         const nestedListType: 'ordered' | 'unordered' = markerStyle.toLowerCase() === 'decimal' ? 'ordered' : 'unordered';
         
@@ -225,7 +209,7 @@ export class XamlListProcessor {
     for (const element of content) {
       if (element && typeof element === 'object') {
         // Use the main converter's processElement method for consistency
-        result += this.converter.processElement(element);
+        result += this.processor.processElement(element);
       }
     }
